+++ src/commands/team.ts
@@ -1,5 +1,5 @@
 import { Command } from "@cliffy/command";
-import { openCommand } from "./team-open.ts";
+
 import { idCommand } from "./team-id.ts";
 import { autolinksCommand } from "./team-autolinks.ts";
 
@@ -8,6 +8,5 @@ export const teamCommand = new Command()
   .action(function () {
     this.showHelp();
   })
-  .command("open", openCommand)
   .command("id", idCommand)
   .command("autolinks", autolinksCommand);