+++ src/utils/actions.ts
@@ -1,12 +1,10 @@
 import { open } from "@opensrc/deno-open";
-import { encodeBase64 } from "@std/encoding/base64";
 import { Select } from "@cliffy/prompt";
 import { branchExists } from "./git.ts";
 import {
   fetchIssueDetails,
   getIssueId,
   getStartedState,
-  getTeamId,
   updateIssueState,
 } from "./linear.ts";
 import { getOption } from "../config.ts";
@@ -34,32 +32,6 @@ export async function openIssuePage(
   const url = `https://linear.app/${workspace}/issue/${issueId}`;
   const destination = options.app ? "Linear.app" : "web browser";
   console.log(`Opening ${url} in ${destination}`);
-  await open(url, options.app ? { app: { name: "Linear" } } : undefined);
-}
-
-export async function openTeamPage(options: { app?: boolean } = {}) {
-  const teamId = await getTeamId();
-  if (!teamId) {
-    console.error(
-      "Could not determine team id from configuration or directory name.",
-    );
-    Deno.exit(1);
-  }
-
-  const workspace = getOption("workspace");
-  if (!workspace) {
-    console.error(
-      "workspace is not set via command line, configuration file, or environment.",
-    );
-    Deno.exit(1);
-  }
-
-  const filterObj = {
-    "and": [{ "assignee": { "or": [{ "isMe": { "eq": true } }] } }],
-  };
-  const filter = encodeBase64(JSON.stringify(filterObj)).replace(/=/g, "");
-  const url =
-    `https://linear.app/${workspace}/team/${teamId}/active?filter=${filter}`;
   await open(url, options.app ? { app: { name: "Linear" } } : undefined);
 }